bulk of parser functionality now in __main__
