+++ shade_ms/__main__.py
@@ -1,6 +1,8 @@
 # -*- coding: future_fstrings -*-
 
 import argparse
+import itertools
+import numpy
 import pkg_resources
 
 from . import DEFAULT_CNUM, DEFAULT_NUM_RENDERS
@@ -473,5 +475,181 @@ def cli():
         )
 
     return parser, optimization_opts
+
+
+# -- parser utility functions --
+# flatten append lists generated by argparse.actions
+flatten_list = lambda list_: list(itertools.chain(*[str_.split(',') for str_ in list_]))
+unpack_axis = lambda opts_axis, default: [default] if not opts_axis else flatten_list(opts_axis)
+# validate limits on axes, both min and max limits must be present
+eval_none = lambda arr_1, arr_2: any([(a is None)^(b is None) for a, b in zip(arr_1, arr_2)])
+# -- parser utility functions --
+
+
+def get_list(list_, type_=str, default=None, set_len=-1):
+    """
+        all parameter lists must be the same length as the expected number of plot
+        and can also impose a type such as float (returning None for an empty string)
+
+        parameters
+        ----------
+        list_: list
+            appended list of input options
+        type_: [optional] object
+            desired variable type: e.g. int, float, str
+            type_=str assumed if not specified
+        default: [optional] object
+            appropriate default value for variable
+            default=None assumed if not specified
+        set_len: [optional] int
+            expected number of plots
+
+        returns
+        -------
+        elems: list
+            list of parameter options to plot,
+            with a parameter for each plot to generate
+    """
+    # if empty list, set default
+    if not list_:
+        elems = [default]
+    else:
+        # stick all lists together
+        elems = flatten_list(list_)
+        # convert type
+        elems = list(map(type_, elems))
+    if len(elems) == 1 and set_len > 0:
+        elems = [elems[0]]*set_len
+    return elems
+
+
+def parse_slice_spec(spec):
+    """ check channel slice
+
+        parameters
+        ----------
+        spec: str
+            format [start]:[stop]:[step]
+
+        returns
+        -------
+        slice: object
+            python slice object
+        elems: list
+            slice element ([start, [stop, [step]]])
+    """
+    if spec:
+        spec_elems = [int(x) if x else None for x in spec.split(":", 2)]
+        return slice(*spec_elems), spec_elems
+    else:
+        return slice(None), []
+
+
+def parse_plot_spec(parser, options):
+    """
+        unpacking the plotting input arguments into variables for plotting and
+        making use the specified parameters makes sense for plotting
+
+        parameters
+        ----------
+        parser: object
+            raw ArgumentParser object
+        options: object
+            Namespace object containing parsed attributes
+
+        returns
+        -------
+        xaxes: list
+            xaxis parameter for each figure
+        yaxes: list
+            yaxis parameter for each figure
+        columns: list
+            name of visibility column per figure
+        caxes: list
+            color axis per figure (if specified, else None)
+        aaxes: list
+            blah, blah, blah, ...
+        areds: list
+            blah, blah, blah, ...
+        xmins: list
+            x-range minimum per figure (if specified, else None)
+        xmaxs: list
+            x-range maximum per figure (if specified, else None)
+        ymins: list
+            y-range minimum per figure (if specified, else None)
+        ymaxs: list
+            y-range maximum per figure (if specified, else None)
+        amins: list
+            blah, blah, blah, ...
+        amaxs: list
+            blah, blah, blah, ...
+        cmins: list
+            minimum colormap value per figure (if specified, else None)
+        cmaxs: list
+            maximum colormap value per figure (if specified, else None)
+        cnums: list
+            number of steps in colormap range
+        chanslice: slice object (if channel selection is specified else All)
+            easy selection of specified channels
+    """
+
+    # figure our list of plots to make
+    # list of xaxis options to plot (default ['TIME'])
+    xaxes = unpack_axis(options.xaxis, 'TIME')
+    # list of yaxis associated with xaxes options (default['DATA:amp'])
+    yaxes = unpack_axis(options.yaxis, 'DATA:amp')
+    if len(xaxes) != len(yaxes):
+        parser.error("--xaxis and --yaxis must be given the same number of times")
+
+    # get list of columns and plot limits of the same length
+    param_desc = {
+                 'names': ('vars', 'opts', 'types', 'defaults'),
+                 'formats': ('U15', 'U15', object, object)
+                 }
+    plot_params = numpy.array([
+                              ('columns', 'col', str, ["DATA"]) ,
+                              ('xmins', 'xmin', float, None),
+                              ('xmaxs', 'xmax', float, None),
+                              ('ymins', 'ymin', float, None),
+                              ('ymaxs', 'ymax', float, None),
+                              ('aaxes', 'aaxis', str, None),
+                              ('amins', 'amin', float, None),
+                              ('amaxs', 'amax', float, None),
+                              ('areds', 'ared', str, 'mean'),
+                              ('caxes', 'colour_by', str, None),
+                              ('cmins', 'cmin', float, None),
+                              ('cmaxs', 'cmax', float, None),
+                              ('cnums', 'cnum', int, DEFAULT_CNUM),
+                              ], dtype=param_desc)
+    if not options.col:
+        options.col = ["DATA"]
+    # For all other settings, returns list same length as xaxes,
+    # or throws error if no conformance.
+    for param, name, dtype, default in zip(plot_params['vars'],
+                                           plot_params['opts'],
+                                           plot_params['types'],
+                                           plot_params['defaults']):
+        elems =  get_list(getattr(options, name, None),
+                          type_=dtype,
+                          default=default,
+                          set_len=len(xaxes),
+                          )
+        if len(elems) > 1 and len(elems) != len(xaxes):
+            parser.error(f"--{name} must be given the same number of times as --xaxis, or else just once")
+        #variable is dynamically created and assigned
+        globals()[param] = elems
+
+    # check min/max
+    if eval_none(xmins, xmaxs):
+        parser.error("--xmin/--xmax must be either both set, or neither")
+    if eval_none(ymins, ymaxs):
+        parser.error("--ymin/--ymax must be either both set, or neither")
+    if eval_none(cmins, cmaxs):
+        parser.error("--cmin/--cmax must be either both set, or neither")
+    if eval_none(amins, amaxs):
+        parser.error("--amin/--amax must be either both set, or neither")
+
+    return [xaxes, yaxes, columns, caxes, aaxes, areds,
+            xmins, xmaxs, ymins, ymaxs, amins, amaxs, cmins, cmaxs, cnums]
 
 # -fin-
