+++ shade_ms/main.py
@@ -5,10 +5,8 @@ matplotlib.use('agg')
 
 import dask.diagnostics
 import datetime
-import itertools
 import json
 import logging
-import numpy
 import os
 import re
 import shade_ms
@@ -19,7 +17,7 @@ import warnings
 from contextlib import contextmanager
 
 from . import DEFAULT_CNUM, DEFAULT_NUM_RENDERS
-from . import cli
+from . import cli, parse_plot_spec, parse_slice_spec
 from . import data_plots, data_mappers
 
 from .data_mappers import DataAxis, col_to_label
@@ -32,74 +30,6 @@ from shade_ms import log, separator
 @contextmanager
 def nullcontext(enter_result=None):
     yield enter_result
-
-
-# -- parser utility functions --
-# flatten append lists generated by argparse.actions
-flatten_list = lambda list_: list(itertools.chain(*[str_.split(',') for str_ in list_]))
-unpack_axis = lambda opts_axis, default: [default] if not opts_axis else flatten_list(opts_axis)
-# validate limits on axes, both min and max limits must be present
-eval_none = lambda arr_1, arr_2: any([(a is None)^(b is None) for a, b in zip(arr_1, arr_2)])
-# -- parser utility functions --
-
-
-def get_list(list_, type_=str, default=None, set_len=-1):
-    """
-        all parameter lists must be the same length as the expected number of plot
-        and can also impose a type such as float (returning None for an empty string)
-
-        parameters
-        ----------
-        list_: list
-            appended list of input options
-        type_: [optional] object
-            desired variable type: e.g. int, float, str
-            type_=str assumed if not specified
-        default: [optional] object
-            appropriate default value for variable
-            default=None assumed if not specified
-        set_len: [optional] int
-            expected number of plots
-
-        returns
-        -------
-        elems: list
-            list of parameter options to plot,
-            with a parameter for each plot to generate
-    """
-    # if empty list, set default
-    if not list_:
-        elems = [default]
-    else:
-        # stick all lists together
-        elems = flatten_list(list_)
-        # convert type
-        elems = list(map(type_, elems))
-    if len(elems) == 1 and set_len > 0:
-        elems = [elems[0]]*set_len
-    return elems
-
-
-def parse_slice_spec(spec):
-    """ check channel slice
-
-        parameters
-        ----------
-        spec: str
-            format [start]:[stop]:[step]
-
-        returns
-        -------
-        slice: object
-            python slice object
-        elems: list
-            slice element ([start, [stop, [step]]])
-    """
-    if spec:
-        spec_elems = [int(x) if x else None for x in spec.split(":", 2)]
-        return slice(*spec_elems), spec_elems
-    else:
-        return slice(None), []
 
 
 def main(argv):
@@ -117,6 +47,28 @@ def main(argv):
     options = parser.parse_args(argv)
     if options.debug:
         shade_ms.log_console_handler.setLevel(logging.DEBUG)
+
+    # pass options to shade_ms
+    data_mappers.set_options(options)
+    data_plots.set_options(options)
+
+    # set colormaps
+    cmap = data_plots.get_colormap(options.cmap)
+    bmap = data_plots.get_colormap(options.bmap)
+    dmap = data_plots.get_colormap(options.dmap)
+
+    # figure our list of plots to make
+    # define plot parameters
+    [xaxes, yaxes, columns, caxes, aaxes, areds,
+     xmins, xmaxs, ymins, ymaxs, amins, amaxs, cmins, cmaxs, cnums] = parse_plot_spec(parser,
+                                                                                      options)
+
+
+    # check chan slice
+    try:
+        chanslice, chanslice_spec = parse_slice_spec(options.chan)
+    except ValueError:
+        parser.error(f"invalid selection --{'chan'} {options.chan}")
 
     # issue warning if only a single antenna is specified
     num_ants_warning = False
@@ -139,82 +91,6 @@ def main(argv):
     if num_ants_warning:
         msg = f"Suggested usage: '--baseline {ant_spec}-*' to select all baselines to {ant_spec}."
         warnings.warn(msg, category=UserWarning)
-
-    # pass options to shade_ms
-    data_mappers.set_options(options)
-    data_plots.set_options(options)
-
-    # set colormaps
-    cmap = data_plots.get_colormap(options.cmap)
-    bmap = data_plots.get_colormap(options.bmap)
-    dmap = data_plots.get_colormap(options.dmap)
-
-    # define plot parameters
-    # xaxes, yaxes, columns, xmins, xmaxs, ymins, ymaxs,
-    # aaxes, amins, amaxs, areds, caxes, cmins, cmaxs, cnums
-    # construct_parameters(parser, options)
-
-    # figure our list of plots to make
-    # list of xaxis options to plot (default ['TIME'])
-    xaxes = unpack_axis(options.xaxis, 'TIME')
-    # list of yaxis associated with xaxes options (default['DATA:amp'])
-    yaxes = unpack_axis(options.yaxis, 'DATA:amp')
-    if len(xaxes) != len(yaxes):
-        parser.error("--xaxis and --yaxis must be given the same number of times")
-
-    # get list of columns and plot limits of the same length
-    param_desc = {
-                 'names': ('vars', 'opts', 'types', 'defaults'),
-                 'formats': ('U15', 'U15', object, object)
-                 }
-    plot_params = numpy.array([
-                              ('columns', 'col', str, ["DATA"]) ,
-                              ('xmins', 'xmin', float, None),
-                              ('xmaxs', 'xmax', float, None),
-                              ('ymins', 'ymin', float, None),
-                              ('ymaxs', 'ymax', float, None),
-                              ('aaxes', 'aaxis', str, None),
-                              ('amins', 'amin', float, None),
-                              ('amaxs', 'amax', float, None),
-                              ('areds', 'ared', str, 'mean'),
-                              ('caxes', 'colour_by', str, None),
-                              ('cmins', 'cmin', float, None),
-                              ('cmaxs', 'cmax', float, None),
-                              ('cnums', 'cnum', int, DEFAULT_CNUM),
-                              ], dtype=param_desc)
-    if not options.col:
-        options.col = ["DATA"]
-    # For all other settings, returns list same length as xaxes,
-    # or throws error if no conformance.
-    for param, name, dtype, default in zip(plot_params['vars'],
-                                           plot_params['opts'],
-                                           plot_params['types'],
-                                           plot_params['defaults']):
-        elems =  get_list(getattr(options, name, None),
-                          type_=dtype,
-                          default=default,
-                          set_len=len(xaxes),
-                          )
-        if len(elems) > 1 and len(elems) != len(xaxes):
-            parser.error(f"--{name} must be given the same number of times as --xaxis, or else just once")
-        #variable is dynamically created and assigned
-        globals()[param] = elems
-
-    # check min/max
-    if eval_none(xmins, xmaxs):
-        parser.error("--xmin/--xmax must be either both set, or neither")
-    if eval_none(ymins, ymaxs):
-        parser.error("--ymin/--ymax must be either both set, or neither")
-    if eval_none(cmins, cmaxs):
-        parser.error("--cmin/--cmax must be either both set, or neither")
-    if eval_none(amins, amaxs):
-        parser.error("--amin/--amax must be either both set, or neither")
-
-    # check chan slice
-    try:
-        chanslice, chanslice_spec = parse_slice_spec(options.chan)
-    except ValueError:
-        parser.error(f"invalid selection --{'chan'} {options.chan}")
 
     log.info(" ".join(sys.argv))
